logging functions reenabled
