+++ app/src/data_module.h
@@ -1,5 +1,9 @@
 #pragma once
 
+#define LOG_SAMPLE_RATE_SPS 125
+#define LOG_WRITE_INTERVAL 10      // Write to file every 10 seconds
+#define LOG_BUFFER_LENGTH 1250 + 1 // 125Hz * 10 seconds
+
 struct hpi_computed_data_t {
     int32_t hr;
     uint32_t rr;
